# docs/specs/v0.1/openapi.json
{
  "components": {
    "schemas": {
...
      "applicability": {
        "additionalProperties": false,
        "properties": {
          "activity": {
            "minLength": 1,
            "type": "string"
          },
          "assessment_id": {
            "pattern": "^[a-z][a-z0-9_.-]*$",
            "type": "string"
          },
          "bundle_hash": {
            "pattern": "^[0-9a-f]{64}$",
            "type": "string"
          },
          "client_class": {
            "minLength": 1,
            "type": "string"
          },
          "conclusion": {
            "enum": [
              "in_scope",
              "out_of_scope",
              "unresolved"
            ]
          },
          "jurisdiction": {
            "minLength": 1,
            "type": "string"
          },
          "legal_entity": {
            "minLength": 1,
            "type": "string"
          },
          "product_class": {
            "minLength": 1,
            "type": "string"
          },
          "reason": {
            "minLength": 1,
            "type": "string"
          },
          "regulated_role": {
            "minLength": 1,
            "type": "string"
          },
          "reviewer_id": {
            "pattern": "^[a-z][a-z0-9_.-]*$",
            "type": "string"
          },
          "source_span_ids": {
            "items": {
              "pattern": "^[a-z][a-z0-9_.-]*$",
              "type": "string"
            },
            "type": "array"
          },
          "valid_from": {
            "format": "date-time",
            "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}\\.[0-9]{6}Z$",
            "type": "string"
          },
          "valid_until": {
            "anyOf": [
              {
                "format": "date-time",
                "pattern": "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}\\.[0-9]{6}Z$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "assessment_id",
          "bundle_hash",
          "legal_entity",
          "regulated_role",
          "activity",
          "product_class",
          "client_class",
          "jurisdiction",
          "valid_from",
          "valid_until",
          "conclusion",
          "reason",
          "source_span_ids",
          "reviewer_id"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "LegalMath public/synthetic workbench",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/applicability": {
      "post": {
        "operationId": "applicability_v1_applicability_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/applicability"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Caller lacks the required role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Record not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Revision, idempotency or release conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Resource budget exceeded"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Invalid request or domain record"
          }
        },
        "summary": "Applicability"
      }
    },
    "/v1/bundles": {
      "post": {
        "operationId": "bundle_create_v1_bundles_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/rule-bundle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Caller lacks the required role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Record not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Revision, idempotency or release conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Resource budget exceeded"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Invalid request or domain record"
          }
        },
        "summary": "Bundle Create"
      }
    },
    "/v1/bundles/{bh}": {
      "get": {
        "operationId": "bundle_get_v1_bundles__bh__get",
        "parameters": [
          {
            "in": "path",
            "name": "bh",
            "required": true,
            "schema": {
              "title": "Bh",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Caller lacks the required role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Record not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Revision, idempotency or release conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Resource budget exceeded"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Invalid request or domain record"
          }
        },
        "summary": "Bundle Get"
      }
    },
    "/v1/bundles/{bh}/build-java": {
      "post": {
        "operationId": "build_v1_bundles__bh__build_java_post",
        "parameters": [
          {
            "in": "path",
            "name": "bh",
            "required": true,
            "schema": {
              "title": "Bh",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Build"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Caller lacks the required role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Record not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Revision, idempotency or release conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Resource budget exceeded"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Invalid request or domain record"
          }
        },
        "summary": "Build"
      }
    },
    "/v1/bundles/{bh}/coverage": {
      "post": {
        "operationId": "record_coverage_v1_bundles__bh__coverage_post",
        "parameters": [
          {
            "in": "path",
            "name": "bh",
            "required": true,
            "schema": {
              "title": "Bh",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Coverage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Caller lacks the required role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Record not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Revision, idempotency or release conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Resource budget exceeded"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Invalid request or domain record"
          }
        },
        "summary": "Record Coverage"
      }
    },
    "/v1/bundles/{bh}/issues/{issue_id}": {
      "post": {
        "operationId": "issue_resolve_v1_bundles__bh__issues__issue_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "bh",
            "required": true,
            "schema": {
              "title": "Bh",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "issue_id",
            "required": true,
            "schema": {
              "title": "Issue Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveIssue"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Caller lacks the required role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Record not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Revision, idempotency or release conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Resource budget exceeded"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Invalid request or domain record"
          }
        },
        "summary": "Issue Resolve"
      }
    },
    "/v1/bundles/{bh}/release": {
      "post": {
        "operationId": "release_v1_bundles__bh__release_post",
        "parameters": [
          {
            "in": "path",
            "name": "bh",
            "required": true,
            "schema": {
              "title": "Bh",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Release"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Caller lacks the required role"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Record not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Revision, idempotency or release conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Resource budget exceeded"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceError"
                }
              }
            },
            "description": "Invalid request or domain record"
          }
        },
        "summary": "Release"
      }
    },
...
  }
}